11-personal-assistant-agent/requirements.txt
openai>=1.50.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
schedule>=1.2.0
